Module oak_containers_stage1::dice

Structs

MemoryRange 🔒

SensitiveDiceDataMemory

Holds a reference to the DICE data in physical memory.

Constants

MEMMAP_PATH 🔒

The path for reading the memory map from the sysfs pseudo-filesystem.

PHYS_MEM_PATH 🔒

The path for reading the physical memory from the mem device.

RESERVED_E820_TYPE 🔒

The expected string representation for reserved memory.

Functions

read_memory_ranges 🔒

Reads the memory ranges which supplied by the firmware to the Linux kernel.