Module oak_containers_stage1::dice

Modules

proto 🔒

Structs

DiceDataBuilder

Builds the DICE evidence and certificate authority for the next DICE layer.

MemoryRange 🔒

Constants

AMD_SEV_SNP_ATTESTATION_REPORT_SIZE 🔒

The size of the attestation report generated by AMD SEV-SNP.

EMPTY_KEY 🔒

A key field made up of all zeroes.

EXPECTED_E820_TYPE 🔒

The expected string representation of the custom type for the reserved memory range that contains the DICE data.

MEMMAP_PATH 🔒

The path for reading the memory map from the sysfs pseudo-filesystem.

P256_PRIVATE_KEY_SIZE 🔒

The actual size used when encoding a Nist P256 private key.

PHYS_MEM_PATH 🔒

The path for reading the physical memory from the mem device.

Functions

cbor_encoded_bytes_to_vec 🔒

Extracts the bytes used to encode a CBOR object from a slice that might include unused bytes by deserializing and re-serializing the encoded CBOR object.

dice_data_to_proto 🔒

Converts the DICE evidence received from Stage 0 to a protocol buffer representation.

extract_stage0_evidence_and_key

Extracts the DICE evidence and ECA key from the Stage 0 DICE data located at the given physical address.

read_memory_ranges 🔒

Reads the memory ranges which supplied by the firmware to the Linux kernel.

read_stage0_dice_data 🔒

Reads the DICE data from the physical memory range starting at the given address.